chore: rename NodeActionModuleInfo
